moved the live response example object to a save space - coverage now 87%!
